split gc into go1.4 and go1.5 variants
